ATM withdrawal
Vinh works for an ATM machine manufacturing company. The basic functionality of an ATM machine
is cash withdrawal. When a user requests a cash withdrawal of W VND (Vietnamese Dong), the ATM
has to dispense N money notes such that they sum up to W . For the next generation of ATM machine,
Vinh is working on an algorithm to minimize the number N of money notes for each cash withdrawal
transaction.
Your task is to help Vinh to do his job given that the money notes come in the values of
1000; 2000; 3000; 5000; 1000·101; 2000·101; 3000·101; 5000·101; : : : ; 1000·10c; 2000·10c; 3000·10c; 5000·10c
where c is a positive integer and Vinh has unlimited supply of money notes for each value.