GOLD (double queue)
The Kingdom ALPHA has n warehouses of golds located on a straight line and are numbered 1, 2,
..., n. The warehouse i has amount of ai (ai is non-negative integer) and is located at coordinate i
(i = 1.. n). The King of ALPHA opens a competition for hunters who are responsible to find a
subset of gold warehouses having largest total amount of golds with respect to the condition that the
distance between two selected warehouses must be greater than or equal to L1 and less than or equal to
L2.