Babylon Tower
Perhaps you have heard of the legend of the Tower of Babylon. Nowadays many details of this tale have
been forgotten. So now, in line with the educational nature of this contest, we will tell you the whole
story:
The babylonians had n types of blocks, and an unlimited supply of blocks of each type. Each type-i block
was a rectangular solid with linear dimensions (xi; yi; zi) . A block could be reoriented so that any two of
its three dimensions determined the dimensions of the base and the other dimension was the height. They
wanted to construct the tallest tower possible by stacking blocks. The problem was that, in building a
tower, one block could only be placed on top of another block as long as the two base dimensions of the
upper block were both strictly smaller than the corresponding base dimensions of the lower block. This
meant, for example, that blocks oriented to have equal-sized bases couldn’t be stacked.
Your job is to write a program that determines the height of the tallest tower the babylonians can build
with a given set of blocks.